web/requirements.txt
streamlit==1.3.0
python-dotenv==0.19.2
bokeh==2.4.1